+++ docs/components/core/PaginationExample.tsx
@@ -1,0 +1,7 @@
+import { Pagination } from "@w-design/core";
+
+const PaginationExample = () => {
+  return <Pagination pages={[1, 2, 3, 4, 5, 6, 7, 8, 9, 10]} />;
+};
+
+export default PaginationExample;