+++ packages/core/src/Button/ButtonGroup.tsx
@@ -3,22 +3,21 @@
  */
 import * as React from "react";
 import styled from "styled-components";
+import { InlineGroupItemStyle } from "../common/generic";
 import Box from "../Box";
 
 const GroupItem = styled.div`
-  flex: 1 0 auto;
-  display: flex;
-  & + &::before {
-    content: "";
-    display: inline-block;
-    width: 0.4rem;
-  }
+  ${InlineGroupItemStyle}
 `;
 
 export const ButtonGroup: React.FC<React.ReactNode> = ({ children }) => (
   <Box display="inline-flex">
     {React.Children.map(children, (child, idx) => {
-      return <GroupItem key={idx}>{React.cloneElement(child as JSX.Element)}</GroupItem>;
+      return (
+        <GroupItem key={idx}>
+          {React.cloneElement(child as JSX.Element)}
+        </GroupItem>
+      );
     })}
   </Box>
 );