+++ packages/core/src/index.ts
@@ -52,6 +52,8 @@ export { default as Steps } from "./Steps";
 
 export { default as Tabs, Tab } from "./Tabs";
 export { default as Typo } from "./Typo";
+
+export * from "./PushMessage";
 export {
   default as PushMessage,
   PushMessageProps,