+++ packages/form/src/Checkbox/index.tsx
@@ -5,7 +5,7 @@ import {
   GlobProps,
   InputControlProps,
   SyntheticEventProps,
-  MouseEventProps
+  MouseEventProps,
 } from "@w-design/core/lib/types/common/props";
 
 export interface CheckboxProps
@@ -31,13 +31,13 @@ const Label = styled.label<CheckboxProps>`
   align-items: flex-start;
   position: relative;
   line-height: 1.2;
-  margin: ${props => props.theme.space[1]}px 0;
+  margin: ${(props) => props.theme.space[1]}px 0;
   user-select: none;
   &:hover {
-    cursor: ${props => (props.isDisabled ? `not-allowed` : `pointer`)};
+    cursor: ${(props) => (props.isDisabled ? `not-allowed` : `pointer`)};
   }
 
-  ${props => props.isDisabled && `opacity: 0.7`};
+  ${(props) => props.isDisabled && `opacity: 0.7`};
 `;
 Label.displayName = "Label";
 
@@ -95,7 +95,6 @@ const CheckboxIcon = styled.span`
 `;
 CheckboxIcon.displayName = "CheckboxIcon";
 
-// TODO: nested checkboxes
 const Checkbox = ({
   checkboxRef,
   name,
@@ -140,9 +139,22 @@ const Checkbox = ({
           {...rest}
         />
         <CheckboxIcon isChecked={isChecked}>
-          <svg width="28" height="28" viewBox="0 0 24 24" focusable="false" role="presentation">
+          <svg
+            width="28"
+            height="28"
+            viewBox="0 0 24 24"
+            focusable="false"
+            role="presentation"
+          >
             <g fillRule="evenodd">
-              <rect fill="currentColor" x="6" y="6" width="12" height="12" rx="2" />
+              <rect
+                fill="currentColor"
+                x="6"
+                y="6"
+                width="12"
+                height="12"
+                rx="2"
+              />
               <path
                 d="M9.707 11.293a1 1 0 1 0-1.414 1.414l2 2a1 1 0 0 0 1.414 0l4-4a1 1 0 1 0-1.414-1.414L11 12.586l-1.293-1.293z"
                 fill="inherit"